requirements.txt
numpy
pillow
ipykernel
opencv-python
mss
pyautogui
statistics
pandas
tensorflow==2.18.0
scikit-learn
